frontend/requirements.txt
streamlit==1.40.2
requests==2.32.3
sseclient-py==1.8.0
python-dotenv==1.0.1
